Move code map fetch to generateAgent
