+++ workspaces/ballerina/ballerina-extension/src/features/ai/executors/base/AICommandExecutor.ts
@@ -102,6 +102,8 @@ export interface AICommandConfig<TParams = any> {
      * summaries to `.ballerina-ai-migration/debug.log` via this logger.
      */
     debugLogger?: MigrationDebugLogger;
+    /** Code map markdown captured at query submission time */
+    codeMapMarkdown?: string;
 }
 
 /**
